+++ NEWS.md
@@ -10,8 +10,14 @@ TODO:
   - e.g. independent test for different omics data and then combine results
 -->
 
+# clusterProfiler 4.9.2
 
-# clusterProfiler 4.9.0.002
++ `append_kegg_category()` function to add KEGG pathway category information to KEGG enrichment result and now it is the default behavior of `enrichKEGG()` and `gseKEGG()` (2023-07-12, Wed)
++ parse KEGG Pathway Category information (2023-07-11, Tue)
++ mv `parse_gff()` to `GOSemSim::read.gaf()` and re-export (2023-07-10, Mon)
++ mv `buildGOmap()` to `GOSemSim::buildGOmap() and re-export
+
+# clusterProfiler 4.9.1
 
 + `getPPI()` to query PPI network from 'stringdb' (2023-05-15, Mon)
 + `getTaxID()` and `getTaxInfo()` functions to query taxonomy information (2023-05-14, Sun)